==> notebooks/0_create_env_file.ipynb <==
# Create Environment Variables File for PoPS Global Model

import os
from getpass import getpass

## Name your configuration scenario

# What should this model scenario be called
short_name = "slf"

# What additional description defines this scenario
# e.g., specific parameter/value of interest + commodity range
add_descript = "emsemble"

sim_name = f"{short_name}_{add_descript}"


## Provide API keys for data acquisition

# Input UN Comtrade premium API key if required
comtrade_auth_key = getpass()

# If notebook was launched from notebook folder of the clone GitHub
# repository, then set working directory to level above
# (e.g., '..' to navigate to /PoPS-Global)
repo_path = '..'
os.chdir(repo_path)
print(os.getcwd())

## Define file paths where data will be loaded and saved

# Provide file paths to where data will be saved
project_dir = str(input())

# Create directories for raw data, inputs, and outputs
# Path to case-study specific raw data inputs (host map, phytosanitary capacity dataset)
data_dir = project_dir + "/Data/"

# Path to formatted model inputs
input_dir = project_dir + "/inputs/"

# Path to save outputs
out_dir = project_dir + "/outputs/"


# Confirm all directories exist:
dir_list = [data_dir, input_dir, out_dir]
for d in dir_list:
    if not os.path.exists(d):
        print(f"ERROR:\t{d} does not exist, creating directory...")
        os.makedirs(d)
    else:
        print(f"Confirmed:\t{d}")

with open(".env", "w") as f:
    f.write(f"COMTRADE_AUTH_KEY={comtrade_auth_key}\n")
    f.write(f"PROJECT_PATH='{project_dir}'\n")
    f.write(f"DATA_PATH='{data_dir}'\n")
    f.write(f"INPUT_PATH='{input_dir}'\n")
    f.write(f"OUTPUT_PATH='{out_dir}'\n")
    f.write(f"SIM_NAME='{sim_name}'\n")
    f.close()
