==> notebooks/2_create_model_config.ipynb <==
# Generate a Configuration File to Run the PoPS Global Model

This notebook creates a configuration file to run a particular model scenario. Some parameters will be fixed across
your case study. Others (calibrated parameters: alpha, lamda, beta, start_year) can be set as a range to be calibrated
and later sampled to run the complete forecasts.

## Imports

import os
import json
import dotenv

## Load Environment Variables and Set Paths

# Navigate one level up to the main repository
os.chdir("..")

# Read environmental variables
env_file = os.path.join(".env")
dotenv.load_dotenv(env_file)

# Path to formatted model inputs
input_dir = os.getenv("INPUT_PATH")

# Path to save outputs
out_dir = os.getenv("OUTPUT_PATH")

# Simulation name
sim_name = os.getenv("SIM_NAME")

config_json_path = f"{out_dir}/config_{sim_name}.json"

if os.path.isfile(config_json_path):
    with open(config_json_path) as file:
        prev_config = json.load(file)
else:
    prev_config = None

## Set Model Parameters
Depending on how you will be running the model, arguments will be provided in different ways.

- **To run the model once** (e.g. to conduct a test run): Set the single sample test-run values (below), and run the
model in notebook 3a (interactive). We recommend running the model interactively first to test that your input data and
model configuration is correct and working.
- **To calibrate the model over a range of possible parameter values** (i.e. to fit key model parameters to the data):
Set range of calibrated parameters (below), and run the model in notebook 3b.  
- **To provide a sample of parameter sets**: After conducting the grid search and evaluating the results (3b), use (3c)
to generate sampled parameter sets and run a forecast of the model. 

You can include the parameters for all three types of runs in this configuration notebook, and the config.json file 
only needs to be created once. However, if you need to modify the parameter values (e.g. to test out a different value, 
make a correction, or expand your grid search parameter range), you should re-run this notebook to upate the config.
json file. 

### Define empty dictionary 
All parameter values will be added to the dictionary, and saved as a JSON file. 

args = {}

### Commodity Data

This information is already included from notebook 1 Data Acquisition. However, you can optionally update the values
here, e.g. to toggle between different trade options that you have already acquired data for. 

# Which temporal resolution, commodities or aggregation should be used
# for historical trade data:

if prev_config:
    timestep = prev_config["timestep"]
    trade_type = prev_config["trade_type"]
    commodity_list = prev_config["commodity_list"]

    print(
        f"Timestep: {timestep}, Trade Type: {trade_type}, Commodity list: {commodity_list}"
    )

# Pulled from Data Aquisition. Option to modify below (e.g. to toggle between sets of already-acquired trade data)
else: 
    timestep = "monthly" # "monthly" or "annual"
    trade_type = "adjusted" # "adjusted" (for individual commodities) or "agg"
    commodity_list = ["6802"]

# Create paths
commodity_data_path = os.path.join(input_dir, f"comtrade/{timestep}_{trade_type}/")

# If no forecast is desired, set argument to None.
commodity_forecast_path = os.path.join(
    input_dir + f"comtrade/trade_forecast/{timestep}_{trade_type}/"
)

# Write to args
args["timestep"] = timestep  # "monthly" or "annual"
args["trade_type"] = trade_type  # "adjusted" (for individual commodities) or "agg"
args["commodity_list"] = commodity_list

args["commodity_path"] = commodity_data_path
args["commodity_forecast_path"] = commodity_forecast_path

# Option to include the path to a lamda weights .csv

args["lamda_weights_path"] = None

# For accessing input file names

if prev_config:
    climate_match_type = prev_config["climate_match"]
    host_only = prev_config["climate_match_host_only"]
    mask = prev_config["mask"]
    threshold_val = prev_config["threshold_val"]
    scaled_min = prev_config["scaled_min"]
    scaled_max = prev_config["scaled_max"]

    print(
        f"Climate match type: {climate_match_type}, Host only?: {host_only}, Mask: {mask},"
        f"Threshold value: {threshold_val}, Scaled Min: {scaled_min}, Scaled Max: {scaled_max}"
    )

# Pulled from Data Aquisition. Option to modify below (e.g. if someone else generated the input files
# or if you already ran Notebook 1 but lost your config.json file).

else: 
    climate_match_type = "origins"  # "origins" or "pairs"
    host_only = False
    mask = "hii"
    threshold_val = 16
    scaled_min = 0.3
    scaled_max = 0.8

# Write to args
args["climate_match"] = climate_match_type
args["climate_match_host_only"] = host_only
args["mask"] = mask
args["threshold_val"] = threshold_val

# Phytosantitary capacity scales
args["scaled_min"] = scaled_min
args["scaled_max"] = scaled_max

### Pest Native to Which Countries?

This information is already included from notebook 1 Data Acquisition.

# Countries where the pest is native or present at first time step of the model run

if prev_config:
    args["native_countries_list"] = prev_config["native_countries_list"]
    print(f'Native countries list: {args["native_countries_list"]}')

# Pulled from Data Aquisition. Option to modify below (NOT recommended)
else:
    args["native_countries_list"] = ["CHN", "VNM"]

### During which months can the pest be present in the shipment?

# List of months when pest can be transported
args["season_dict"] = {
    "NH_season": ["09", "10", "11", "12", "01", "02", "03", "04"],
    "SH_season": ["04", "05", "06", "07", "08", "09", "10"],
}

### Model Equation Parameter Values - Single Run 

Define the parameter values for the model equation, to be used in a single test run. 

# Define test-run parameter values (single sample)

args["alpha"] = 0.7
args["beta"] = 0.5
args["lamda_c_list"] = [0.8]
args["start_year"] = 2005


# Confirm that the length of lambda_c_list matches that of the commodity_list

print("Commodity match check:")

if len(args["lamda_c_list"]) == len(args["commodity_list"]):
    print("The length of the commodity list and the commodity parameter (lamda) match!")
else:
    print(
        "OOPS! The length of the commodity list and commodity parameter (lamda) do not match.\n"
        "Please correct either to ensure that the model will be able to run."
    )

### Model Equation Parameter Values - Calibration Runs

Define the parameter values for the model equation, to be used in a calibration grid-search. 

# Define grid-search range of calibrated parameter values

args["alphas"] = [0.25, 0.5, 0.75]
args["betas"] = [0.5]
args["lamdas"] = [0.8, 0.9, 1, 1.1]

args["start_years"] = [2005, 2006]

# Run counting is inclusive of the start and end run 

args["start_run"] = 1
args["end_run"] = 80 

# Set cross-validation method (optional)
# "loo" for leave-one-out cross validation
# None for standard calibration

args["validation_method"] = "loo"

### Model Equation Parameter Values - Not Calibrated 

Define the parameter values for the model equation that will be static across test runs, calibration, and forecasts. 

# Define static parameter values

args["mu"] = 0.0
args["phi"] = 1
args["w_phi"] = 1
args["sigma_epsilon"] = 0.5
args["sigma_phi"] = 1
args["stop_year"] = 2019

# Set random seed (optional)

args["random_seed"] = None

Define how transmission lags occur. Stochastic lags are drawn from a gamma distribution generated from shape and scale parameters.

# Define transmission lag values

args["time_to_infectivity"] = None

args["transmission_lag_type"] = "stochastic"
args["gamma_shape"] = 4
args["gamma_scale"] = 1

Define what outputs will be written to file. By default, a .csv of origin-destination pairs for all introductions, 
as well as the aggregated probability of introduction to each country at each time step are output. 

You can optionally export additional granularity by setting the below options to `True`, however it will increase 
time to write run outputs (and the number of files generated).

# Save n x n matrices for each time step where n is the number of countries,
# and values represent the origin-destination probability of entry or
# probability of establishment

args["save_entry"] = False
args["save_estab"] = False
args["save_intro"] = False
args["save_country_intros"] = False

### Define scenarios (optional)

Scenarios describe how changes are made to trade, to reflect potential management actions. 

scenario_list = []

# for i in range(2010, 2030):
#     start_scenario = [2010, 'CHN', 'USA', 'decrease', 1]
#     new_scenario = start_scenario
#     new_scenario[0] = i
#     scenario_list .append(new_scenario)

# for i in range(2014, 2030):
#     start_scenario = [2014, 'JPN', 'USA', 'decrease', 0.8]
#     new_scenario = start_scenario
#     new_scenario[0] = i
#     scenario_list .append(new_scenario)

# for i in range(2014, 2030):
#     start_scenario = [2014, 'KOR', 'USA', 'decrease', 0.8]
#     new_scenario = start_scenario
#     new_scenario[0] = i
#     scenario_list .append(new_scenario)

# for i in range(2020, 2030):
#     start_scenario = [2020, 'ITA', 'USA', 'decrease', 0.8]
#     new_scenario = start_scenario
#     new_scenario[0] = i
#     scenario_list .append(new_scenario)

# for i in range(2020, 2030):
#     start_scenario = [2020, 'TUR', 'USA', 'decrease', 0.8]
#     new_scenario = start_scenario
#     new_scenario[0] = i
#     scenario_list .append(new_scenario)


args["scenario_list"] = scenario_list

### Configure summary statistics

Summary statistics are calculated on the combined output across multiple runs. You can define several
parameters (described below) that will adjust how certain statistics are calculated.

# What is your primary country of interest? (format = ISO3 code)
args["coi"] = "USA"

# Calculate cumulative probability of intro to COI for the following years:

args["sim_years"] = [2014, 2020]

# End-valid year: At what year do you want to stop evaluating the summary statistics?
# E.g. often, present year or year near the last known introduction

args["end_valid_year"] = 2019

# How many years BEFORE and AFTER the first record do you consider a presence to be accurate?

# BEFORE: e.g. Model predicts 2013 introduction, but first documented record was 2015. This may still be
# considered accurate if you (1) consider that first documented records may have a detection lag
# from the true first introduction date and (2) if you prefer to favor risk-conservative forecasts
# (avoid penalizing reasonably early predictions that are useful for preventative measures)

args["years_before_firstRecord"] = 4

# AFTER: e.g. Model predicts 2016 introduction, but first documented record was 2015. This may be considered
# accurate if you are less concerned with capturing the exact temporal window, but could be considered
# not accurate if you are mainly concerned with preventative measures (set to 0).

args["years_after_firstRecord"] = 0

### Cores to use for parallel processing

The model calibration and forecast can be run in parallel, on multiple CPUs. You can control the resources the 
model will use with the below parameter to either speed up the runs or prevent the model from running too many
processes on your computer (e.g. to limit usage, set to 1; to increase speed, set as high as you have CPUs 
available).

args["cores_to_use"] = 4

# Not sure how many CPUs to use? If you won't be running anything else in the meantime, you can use
# most of your CPUs

print(
    f"Your computer has {os.cpu_count()} CPUs. "
    f"You can safely set 'cores_to_use' as high as {os.cpu_count() - 1}."
    )

## Write and save configuration file

if os.path.isfile(config_json_path):
    with open(config_json_path) as file:
        prev_config = json.load(file)

    prev_config.update(args)

    with open(config_json_path, mode="w") as f:
        f.write(json.dumps(prev_config, indent=4))

else:
    with open(config_json_path, "w") as file:
        json.dump(args, file, indent=4)

print("\tSaved ", config_json_path)

## Next: Run model (interactive)

Run the model first interactively to test that your data is complete, parameters are sensible, and the model 
is configured to run (Notebook 3A). Then, go on to calibrate (3B) and forecast (3C) with the model.
